+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon src/index.js"
+    "start": "nodemon src/index.js -e ejs,js"
   },
   "keywords": [
     "UrlShortener"
@@ -13,6 +13,8 @@
   "author": "Group-31",
   "license": "ISC",
   "dependencies": {
+    "dotenv": "^16.0.1",
+    "ejs": "^3.1.8",
     "express": "^4.18.1",
     "mongoose": "^6.3.3",
     "redis": "^3.1.2",
