+++ src/routes.js
@@ -7,3 +7,27 @@ router.post('/url/shorten', urlController.shortenUrl)
 router.get('/:urlCode', urlController.getUrl)
 
 module.exports = router
+
+// app.post("/sign_up",(req,res)=>{
+//     var name = req.body.name;
+//     var email = req.body.email;
+//     var phno = req.body.phno;
+//     var password = req.body.password;
+
+//     var data = {
+//         "name": name,
+//         "email" : email,
+//         "phno": phno,
+//         "password" : password
+//     }
+
+//     db.collection('users').insertOne(data,(err,collection)=>{
+//         if(err){
+//             throw err;
+//         }
+//         console.log("Record Inserted Successfully");
+//     });
+
+//     return res.redirect('form.html')
+
+// })
