+++ views/error.ejs
@@ -1,0 +1,21 @@
+
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title> Shortify</title>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
+        integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+
+    <link rel="stylesheet" type="text/css" href="style.css">
+</head>
+
+<body>
+
+    <h4>
+        <%= Error %>
+    </h4>
+    
+</body>
+
+</html>
